requirements.txt
asn1crypto==1.2.0
astroid==2.3.1
autopep8==1.4.4
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.3.1
Django==2.2.7
django-bootstrap-modal-forms==1.4.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-filter==2.2.0
django-polls==0.1
entrypoints==0.3
flake8==3.7.8
idna==2.8
isort==4.3.21
jupyter-contrib-core==0.3.3
lazy-object-proxy==1.4.2
mccabe==0.6.1
# Pillow==6.1.0
# psycopg2-binary==2.8.3
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.0
python-xmp-toolkit==2.0.1
pytz==2019.1
requests==2.22.0
ruamel-yaml==0.15.46
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
tqdm==4.36.1
typed-ast==1.4.0
urllib3==1.24.2
wincertstore==0.2
wrapt==1.11.2
